refactor: improved the dashboard code quality by breaking it into components
