+++ apps/dashboard/App.tsx
@@ -1,195 +1,19 @@
-import { useCallback, useEffect, useMemo, useRef, useState } from 'react'
-import { cn } from './lib/utils'
-import type { CommentRecord, ShareEventsResponse } from './api'
+import { useCallback, useEffect, useMemo, useState } from 'react'
 import { updateImplementationStatus as apiUpdateImpl, updateReviewStatus as apiUpdateReview } from './api'
 import { useProjects } from './hooks/useProjects'
 import { useComments } from './hooks/useComments'
-import { useAgentSession, type PromptTarget } from './hooks/useAgentSession'
+import { useAgentSession } from './hooks/useAgentSession'
+import { getDisplayStatus, isInactive, mapServerComment } from './lib/comment'
+import { AGENTS, type Comment, type ImplStatus, type ReviewStatus, type StatusFilter } from './lib/types'
+import { Header } from './components/Header'
+import { CommentList } from './components/CommentList'
+import { CommentDetail } from './components/CommentDetail'
+import { AgentSidebar } from './components/AgentSidebar'
+import { StatusBar } from './components/StatusBar'
+import { CommandPalette } from './components/CommandPalette'
 
 const API_BASE = import.meta.env.VITE_API_BASE || '/api'
 const REVIEWER_TOKEN = import.meta.env.VITE_REVIEWER_TOKEN || ''
-
-// ─── Types ──────────────────────────────────────────────────────────
-
-type ReviewStatus = 'open' | 'accepted' | 'rejected'
-type ImplStatus = 'unassigned' | 'claimed' | 'in_progress' | 'blocked' | 'done'
-type StatusFilter = 'all' | 'open' | 'ready' | 'done'
-type DisplayStatus = 'open' | 'ready' | 'done' | 'rejected'
-
-interface Comment {
-  id: string
-  projectId: string
-  pageUrl: string
-  selector: string
-  x: number
-  y: number
-  body: string
-  reviewStatus: ReviewStatus
-  implementationStatus: ImplStatus
-  claimedByAgentId: string | null
-  createdAt: string
-  updatedAt: string
-  // Mock-only fields
-  author: string
-  authorInitial: string
-  authorColor: string
-  screenshotUrl: string | null
-}
-
-// ─── Static Data ─────────────────────────────────────────────────────
-
-const AGENTS: Array<{ id: string; name: string; hint: string; target: PromptTarget }> = [
-  { id: 'claude-code', name: 'Claude Code', hint: 'Paste prompt in chat', target: 'claude-code' },
-  { id: 'codex', name: 'Codex', hint: 'Paste prompt in prompt', target: 'codex' },
-  { id: 'cursor', name: 'Cursor', hint: 'Paste prompt in composer', target: 'generic' },
-  { id: 'windsurf', name: 'Windsurf', hint: 'Paste prompt in Cascade', target: 'generic' },
-  { id: 'cline', name: 'Cline', hint: 'Paste prompt in chat', target: 'generic' },
-]
-
-const AUTHOR_COLORS = ['#6366F1', '#8B5CF6', '#EC4899', '#F59E0B', '#10B981', '#0EA5E9']
-
-type FeedbackEvent = ShareEventsResponse['events'][number]
-
-function describeEvent(ev: FeedbackEvent): { kind: 'done' | 'claim' | 'file' | 'other'; text: string } {
-  const actor = ev.actorId || 'agent'
-  const summary = typeof ev.payload?.summary === 'string' ? ev.payload.summary as string : null
-  const note = typeof ev.payload?.note === 'string' ? ev.payload.note as string : null
-  switch (ev.eventType) {
-    case 'comment.claimed':
-      return { kind: 'claim', text: `${actor} claimed comment` }
-    case 'comment.started':
-      return { kind: 'file', text: `${actor} started${summary ? ` · ${summary}` : ''}` }
-    case 'comment.noted':
-      return { kind: 'other', text: `${actor} note${note ? ` · ${note}` : ''}` }
-    case 'comment.blocked':
-      return { kind: 'other', text: `${actor} blocked${note ? ` · ${note}` : ''}` }
-    case 'comment.completed':
-      return { kind: 'done', text: `${actor} done${summary ? ` · ${summary}` : ''}` }
-    case 'comment.reopened':
-      return { kind: 'other', text: `${actor} reopened comment` }
-    case 'comment.reviewed':
-      return { kind: 'other', text: `Reviewer marked ${ev.payload?.reviewStatus ?? 'updated'}` }
-    case 'comment.implementation_changed':
-      return { kind: 'other', text: `Status → ${ev.payload?.implementationStatus ?? 'updated'}` }
-    case 'presence.updated':
-      return { kind: 'other', text: `${actor} ${typeof ev.payload?.status === 'string' ? ev.payload.status as string : 'present'}${summary ? ` · ${summary}` : ''}` }
-    case 'share.created':
-      return { kind: 'other', text: `Session opened` }
-    default:
-      return { kind: 'other', text: ev.eventType }
-  }
-}
-
-// ─── Helpers ────────────────────────────────────────────────────────
-
-function timeAgo(iso: string) {
-  const s = Math.max(0, Math.floor((Date.now() - new Date(iso).getTime()) / 1000))
-  if (s < 60) return 'just now'
-  if (s < 3600) return `${Math.floor(s / 60)}m ago`
-  if (s < 86400) return `${Math.floor(s / 3600)}h ago`
-  return `${Math.floor(s / 86400)}d ago`
-}
-
-function truncateUrl(url: string) {
-  return url.startsWith('/') ? url : url.replace(/^https?:\/\/[^/]+/, '')
-}
-
-function formatDocUrl(url: string) {
-  try {
-    const u = new URL(url)
-    const slug = u.searchParams.get('fw_share') ?? ''
-    const token = u.searchParams.get('token') ?? ''
-    const tokenShort = token.length > 8 ? `${token.slice(0, 6)}…${token.slice(-3)}` : token
-    return `${u.host}/?fw_share=${slug}&token=${tokenShort}`
-  } catch {
-    return url
-  }
-}
-
-function buildIntegrationPrompt(projectId: string, apiBase: string) {
-  return `Install the Design Project feedback widget so reviewers can leave visual feedback in this app.
-
-1. Install the package:
-
-   npm install @thedesignproject/feedback-widget
-
-2. Mount <FeedbackWidget /> in the root React component (e.g. App.tsx or a top-level layout) so it appears on every page:
-
-   import { FeedbackWidget } from '@thedesignproject/feedback-widget'
-
-   export default function App() {
-     return (
-       <>
-         {/* existing app */}
-         <FeedbackWidget
-           apiBase="${apiBase}"
-           projectId="${projectId}"
-         />
-       </>
-     )
-   }
-
-The widget injects a floating pin tool. Reviewers click anywhere on the page to drop a pin and attach a comment + screenshot, which appear in this dashboard.
-
-Use these exact config values:
-- apiBase:   ${apiBase}
-- projectId: ${projectId}
-
-Follow the codebase's existing conventions (TypeScript, formatter, project structure). Don't refactor unrelated code.`
-}
-
-function hashString(s: string) {
-  let h = 0
-  for (let i = 0; i < s.length; i++) h = (h * 31 + s.charCodeAt(i)) >>> 0
-  return h
-}
-
-function authorColorFor(name: string) {
-  return AUTHOR_COLORS[hashString(name) % AUTHOR_COLORS.length]
-}
-
-function mapServerComment(record: CommentRecord): Comment {
-  const author = record.authorName?.trim() || 'Anonymous'
-  const initial = author.charAt(0).toUpperCase() || '?'
-  return {
-    id: record.id,
-    projectId: record.projectId,
-    pageUrl: record.pageUrl,
-    selector: record.selector || '',
-    x: record.x,
-    y: record.y,
-    body: record.body,
-    reviewStatus: record.reviewStatus,
-    implementationStatus: record.implementationStatus,
-    claimedByAgentId: record.claimedByAgentId,
-    createdAt: record.createdAt,
-    updatedAt: record.updatedAt,
-    author,
-    authorInitial: initial,
-    authorColor: authorColorFor(author),
-    screenshotUrl: record.imageUrl,
-  }
-}
-
-function isInactive(c: Comment) {
-  return c.reviewStatus === 'rejected' || c.implementationStatus === 'done'
-}
-
-function getDisplayStatus(c: Comment): DisplayStatus {
-  if (c.reviewStatus === 'rejected') return 'rejected'
-  if (c.implementationStatus === 'done') return 'done'
-  if (c.reviewStatus === 'accepted') return 'ready'
-  return 'open'
-}
-
-const DISPLAY_STATUS_LABELS: Record<DisplayStatus, string> = {
-  open: 'Open',
-  ready: 'Ready for Agent',
-  done: 'Done',
-  rejected: 'Rejected',
-}
-
-// ─── App ────────────────────────────────────────────────────────────
 
 export function App() {
   const { projects, loading: projectsLoading, error: projectsError, createProject } = useProjects(API_BASE, REVIEWER_TOKEN)
@@ -450,663 +274,82 @@ export function App() {
 
   return (
     <div className="flex flex-col h-screen overflow-hidden">
+      <Header
+        projects={projects}
+        projectsLoading={projectsLoading}
+        projectsError={projectsError}
+        commentsLoading={commentsLoading}
+        selectedProject={selectedProject}
+        commentCount={comments.length}
+        setSelectedProject={setSelectedProject}
+        setStatusFilter={setStatusFilter}
+        setSelectedCommentId={setSelectedCommentId}
+        addProjectOpen={addProjectOpen}
+        setAddProjectOpen={setAddProjectOpen}
+        onAddProject={handleAddProject}
+        addProjectBusy={addProjectBusy}
+        addProjectError={addProjectError}
+        onOpenCmd={() => setCmdOpen(true)}
+        theme={theme}
+        toggleTheme={() => setTheme((t) => t === 'light' ? 'dark' : 'light')}
+      />
 
-      {/* ── Top Header ── */}
-      <header className="flex items-center gap-3 px-5 h-[52px] shrink-0 border-b border-border bg-card">
-        <div className="flex items-center gap-2 mr-2">
-          <div className="w-6 h-6 rounded-md bg-primary flex items-center justify-center">
-            <svg width="13" height="13" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2.5" strokeLinecap="round" strokeLinejoin="round" className="text-primary-foreground">
-              <path d="M21 15a2 2 0 01-2 2H7l-4 4V5a2 2 0 012-2h14a2 2 0 012 2v10z" />
-            </svg>
-          </div>
-          <span className="text-sm font-bold tracking-tight text-foreground font-serif italic">feedback</span>
-        </div>
-
-        <div className="w-px h-5 bg-border" />
-
-        <nav className="flex items-center gap-1 flex-1 overflow-auto">
-          {projectsLoading && projects.length === 0 ? (
-            <span className="text-xs text-muted-foreground px-2">Loading projects…</span>
-          ) : projectsError ? (
-            <span className="text-xs text-status-rejected px-2">{projectsError}</span>
-          ) : projects.length === 0 ? (
-            <span className="text-xs text-muted-foreground px-2">No projects yet</span>
-          ) : (
-            projects.map((p) => {
-              const isActive = selectedProject === p.publicKey
-              const count = isActive ? comments.length : null
-              return (
-                <button
-                  key={p.publicKey}
-                  onClick={() => { setSelectedProject(p.publicKey); setStatusFilter('all'); setSelectedCommentId('') }}
-                  className={cn(
-                    'px-3 py-1.5 rounded-md text-xs font-semibold transition-all whitespace-nowrap',
-                    isActive
-                      ? 'bg-primary text-primary-foreground'
-                      : 'text-muted-foreground hover:text-foreground hover:bg-accent'
-                  )}
-                >
-                  {p.name}
-                  {isActive && commentsLoading ? (
-                    <span className="ml-1.5 inline-flex items-center">
-                      <Spinner size={11} strokeWidth={3} className="" />
-                    </span>
-                  ) : count !== null && count > 0 ? (
-                    <span className="ml-1.5 text-[10px] font-bold px-1.5 py-0.5 rounded-full bg-primary-foreground/20 text-primary-foreground">
-                      {count}
-                    </span>
-                  ) : null}
-                </button>
-              )
-            })
-          )}
-
-          <div data-add-project className="relative">
-            <button
-              onClick={() => setAddProjectOpen((v) => !v)}
-              className={cn(
-                'w-7 h-7 rounded-md flex items-center justify-center text-muted-foreground transition-all',
-                addProjectOpen
-                  ? 'bg-accent text-foreground'
-                  : 'hover:bg-accent hover:text-foreground'
-              )}
-              title="Add project"
-            >
-              <PlusIcon size={14} />
-            </button>
-            {addProjectOpen && (
-              <AddProjectPopover
-                onAdd={handleAddProject}
-                onClose={() => setAddProjectOpen(false)}
-                busy={addProjectBusy}
-                error={addProjectError}
-              />
-            )}
-          </div>
-        </nav>
-
-        <div className="flex items-center gap-2 ml-auto">
-          <button
-            onClick={() => setCmdOpen(true)}
-            className="flex items-center gap-2 px-3 py-1.5 rounded-md border border-border bg-background text-muted-foreground text-xs w-52 hover:border-muted-foreground/30 hover:bg-accent transition-colors cursor-pointer"
-          >
-            <SearchIcon />
-            <span className="flex-1 text-left">Search Feedback...</span>
-            <kbd className="hidden sm:inline-flex items-center gap-0.5 px-1.5 py-0.5 rounded border border-border bg-card text-[10px] font-mono text-muted-foreground">
-              <span className="text-[11px]">⌘</span>K
-            </kbd>
-          </button>
-          <button
-            onClick={() => setTheme((t) => t === 'light' ? 'dark' : 'light')}
-            title={theme === 'light' ? 'Switch to dark mode' : 'Switch to light mode'}
-            className="w-7 h-7 rounded-md flex items-center justify-center text-muted-foreground hover:text-foreground hover:bg-accent transition-colors"
-          >
-            {theme === 'light' ? (
-              <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
-                <path d="M21 12.79A9 9 0 1 1 11.21 3 7 7 0 0 0 21 12.79z" />
-              </svg>
-            ) : (
-              <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
-                <circle cx="12" cy="12" r="4" />
-                <path d="M12 2v2M12 20v2M4.93 4.93l1.41 1.41M17.66 17.66l1.41 1.41M2 12h2M20 12h2M4.93 19.07l1.41-1.41M17.66 6.34l1.41-1.41" />
-              </svg>
-            )}
-          </button>
-          <div className="w-7 h-7 rounded-full bg-primary/10 flex items-center justify-center text-xs font-bold text-primary">
-            TO
-          </div>
-        </div>
-      </header>
-
-      {/* ── Main 3-panel layout ── */}
       <div className="flex flex-1 overflow-hidden">
+        <CommentList
+          filteredComments={filteredComments}
+          counts={counts}
+          statusFilter={statusFilter}
+          selectFilter={selectFilter}
+          bulkMode={bulkMode}
+          enterBulkMode={() => setBulkMode(true)}
+          exitBulkMode={exitBulkMode}
+          bulkSelectedIds={bulkSelectedIds}
+          toggleSelectAllVisible={toggleSelectAllVisible}
+          applyBulkAction={applyBulkAction}
+          toggleBulkSelect={toggleBulkSelect}
+          commentsLoading={commentsLoading}
+          commentsError={commentsError}
+          selectedCommentId={selectedCommentId}
+          setSelectedCommentId={setSelectedCommentId}
+        />
 
-        {/* ── Left Panel: Comment List ── */}
-        <div className="w-[400px] shrink-0 flex flex-col border-r border-border bg-card">
+        <CommentDetail
+          selectedComment={selectedComment}
+          selectedProject={selectedProject}
+          apiBase={API_BASE}
+          commentsLoading={commentsLoading}
+          commentsError={commentsError}
+          projectComments={projectComments}
+          filteredComments={filteredComments}
+          selectedIdx={selectedIdx}
+          goPrev={goPrev}
+          goNext={goNext}
+          toggleReview={toggleReview}
+          handleToggleDone={handleToggleDone}
+        />
 
-          <div className="px-4 pt-4 pb-2.5 border-b border-border">
-            <div className="flex items-center justify-between mb-3">
-              <h2 className="text-base font-bold text-foreground tracking-tight">
-                {counts.all} Feedback Items
-              </h2>
-              <button
-                onClick={() => bulkMode ? exitBulkMode() : setBulkMode(true)}
-                className={cn(
-                  'text-[11px] font-medium transition-colors flex items-center gap-1',
-                  bulkMode ? 'text-foreground' : 'text-muted-foreground hover:text-foreground'
-                )}
-              >
-                <CheckboxIcon /> {bulkMode ? 'Cancel' : 'Select'}
-              </button>
-            </div>
-            <div className="flex gap-1">
-              {(['all', 'open', 'ready', 'done'] as StatusFilter[]).map((f) => {
-                const count = counts[f as keyof typeof counts] ?? 0
-                const label = f === 'all' ? 'All' : f === 'open' ? 'Open' : f === 'ready' ? 'Ready for Agent' : 'Done'
-                return (
-                  <button
-                    key={f}
-                    onClick={() => selectFilter(f)}
-                    className={cn(
-                      'px-2.5 py-1 rounded-md text-[11px] font-semibold transition-all',
-                      statusFilter === f
-                        ? 'bg-primary text-primary-foreground'
-                        : 'text-muted-foreground hover:text-foreground hover:bg-accent'
-                    )}
-                  >
-                    {label}
-                    <span className={cn(
-                      'ml-1.5 text-[10px] font-bold min-w-[18px] text-center py-0.5 px-1 rounded-full',
-                      statusFilter === f
-                        ? 'bg-primary-foreground/20 text-primary-foreground'
-                        : 'bg-muted text-muted-foreground/60'
-                    )}>{count}</span>
-                  </button>
-                )
-              })}
-            </div>
-          </div>
-
-          {/* Bulk action bar */}
-          {bulkMode && (
-            <div className="px-4 py-2.5 border-b border-border bg-accent/40 flex items-center gap-2 whitespace-nowrap">
-              <button
-                onClick={toggleSelectAllVisible}
-                className="text-[11px] font-semibold text-muted-foreground hover:text-foreground transition-colors whitespace-nowrap"
-              >
-                {filteredComments.length > 0 && filteredComments.every((c) => bulkSelectedIds.has(c.id)) ? 'Deselect all' : 'Select all'}
-              </button>
-              <span className="text-[11px] text-muted-foreground whitespace-nowrap">
-                {bulkSelectedIds.size} selected
-              </span>
-              <div className="flex-1" />
-              <button
-                disabled={bulkSelectedIds.size === 0}
-                onClick={() => applyBulkAction('ready')}
-                className="text-[11px] font-semibold px-2.5 py-1 rounded-md bg-status-accepted-bg text-status-accepted hover:bg-status-accepted hover:text-white transition-colors disabled:opacity-30 disabled:pointer-events-none whitespace-nowrap"
-              >
-                Ready
-              </button>
-              <button
-                disabled={bulkSelectedIds.size === 0}
-                onClick={() => applyBulkAction('done')}
-                className="text-[11px] font-semibold px-2.5 py-1 rounded-md bg-status-done-bg text-status-done hover:bg-status-done hover:text-white transition-colors disabled:opacity-30 disabled:pointer-events-none whitespace-nowrap"
-              >
-                Done
-              </button>
-              <button
-                disabled={bulkSelectedIds.size === 0}
-                onClick={() => applyBulkAction('reject')}
-                className="text-[11px] font-semibold px-2.5 py-1 rounded-md bg-status-rejected-bg text-status-rejected hover:bg-status-rejected hover:text-white transition-colors disabled:opacity-30 disabled:pointer-events-none whitespace-nowrap"
-              >
-                Reject
-              </button>
-            </div>
-          )}
-
-          <div className="flex-1 overflow-y-auto">
-            {commentsLoading ? (
-              <div className="flex flex-col items-center justify-center h-full text-center px-8 gap-3">
-                <Spinner />
-                <p className="text-xs text-muted-foreground">Loading comments…</p>
-              </div>
-            ) : commentsError ? (
-              <div className="flex flex-col items-center justify-center h-full text-center px-8">
-                <p className="text-sm font-semibold text-status-rejected mb-1">Failed to load</p>
-                <p className="text-xs text-muted-foreground">{commentsError}</p>
-              </div>
-            ) : filteredComments.length === 0 ? (
-              <div className="flex flex-col items-center justify-center h-full text-center px-8">
-                <div className="w-10 h-10 rounded-xl bg-muted flex items-center justify-center mb-3">
-                  <ChatIcon className="text-muted-foreground" />
-                </div>
-                <p className="text-sm font-semibold text-foreground mb-1">No comments</p>
-                <p className="text-xs text-muted-foreground">Nothing here yet for this filter.</p>
-              </div>
-            ) : (
-              <div className="animate-stagger">
-              {filteredComments.map((comment) => {
-                const isActive = comment.id === selectedCommentId
-                const isChecked = bulkSelectedIds.has(comment.id)
-                const inactive = isInactive(comment)
-                const borderColor =
-                  comment.implementationStatus === 'done' ? 'border-l-status-done' :
-                  comment.reviewStatus === 'accepted' ? 'border-l-status-accepted' :
-                  comment.reviewStatus === 'rejected' ? 'border-l-status-rejected' :
-                  'border-l-transparent'
-
-                return (
-                  <button
-                    key={comment.id}
-                    onClick={() => bulkMode ? toggleBulkSelect(comment.id) : setSelectedCommentId(comment.id)}
-                    className={cn(
-                      'w-full text-left px-4 py-3 border-b border-border/50 border-l-[3px] card-hover',
-                      borderColor,
-                      bulkMode && isChecked ? 'bg-primary/10' : isActive && !bulkMode ? 'bg-accent' : 'hover:bg-accent/40',
-                      inactive && !isChecked && 'opacity-50'
-                    )}
-                  >
-                    <div className="flex items-center justify-between mb-1">
-                      <div className="flex items-center gap-2">
-                        {bulkMode ? (
-                          <div className={cn(
-                            'w-[18px] h-[18px] rounded shrink-0 flex items-center justify-center border-2 transition-colors',
-                            isChecked ? 'bg-primary border-primary' : 'border-muted-foreground/40'
-                          )}>
-                            {isChecked && (
-                              <svg width="10" height="10" viewBox="0 0 24 24" fill="none" stroke="white" strokeWidth="3" strokeLinecap="round" strokeLinejoin="round">
-                                <polyline points="20 6 9 17 4 12" />
-                              </svg>
-                            )}
-                          </div>
-                        ) : (
-                          <div
-                            className="w-[18px] h-[18px] rounded-full shrink-0 flex items-center justify-center text-[8px] font-bold text-white"
-                            style={{ background: comment.authorColor }}
-                          >
-                            {comment.authorInitial}
-                          </div>
-                        )}
-                        <span className="text-[13px] font-bold text-foreground">{comment.author}</span>
-                      </div>
-                      <span className="text-[10px] text-muted-foreground/60">{timeAgo(comment.createdAt)}</span>
-                    </div>
-
-                    <p className={cn(
-                      'text-[12px] leading-relaxed mb-2 line-clamp-2 pl-[26px]',
-                      inactive ? 'text-muted-foreground/50 line-through' : 'text-muted-foreground'
-                    )}>
-                      {comment.body}
-                    </p>
-
-                    <div className="flex items-center gap-1.5 pl-[26px] flex-wrap">
-                      <StatusBadge comment={comment} />
-                      {comment.claimedByAgentId && (
-                        <span className="inline-flex items-center gap-1 text-[10px] font-semibold text-status-in-progress">
-                          <span className="w-1.5 h-1.5 rounded-full bg-status-in-progress animate-pulse-dot" />
-                          {comment.claimedByAgentId}
-                        </span>
-                      )}
-                      <span className="text-[10px] text-muted-foreground/40 font-mono ml-auto truncate max-w-[140px]">
-                        {comment.selector.split(' > ').pop()}
-                      </span>
-                    </div>
-                  </button>
-                )
-              })}
-              </div>
-            )}
-          </div>
-        </div>
-
-        {/* ── Center Panel: Comment Detail ── */}
-        <div className="flex-1 flex flex-col overflow-hidden bg-background">
-          {selectedComment ? (
-            <>
-              <div className="flex items-center justify-between px-6 h-[44px] shrink-0 border-b border-border bg-card">
-                <div className="flex items-center gap-2 text-xs text-muted-foreground">
-                  <span className="font-mono font-medium">#{selectedComment.id}</span>
-                  <span>·</span>
-                  <span className="font-mono">{truncateUrl(selectedComment.pageUrl)}</span>
-                  <span>·</span>
-                  {(() => {
-                    const ds = getDisplayStatus(selectedComment)
-                    return (
-                      <span className={cn(
-                        'font-semibold',
-                        ds === 'ready' && 'text-status-accepted',
-                        ds === 'rejected' && 'text-status-rejected',
-                        ds === 'done' && 'text-status-done',
-                        ds === 'open' && 'text-muted-foreground',
-                      )}>
-                        {DISPLAY_STATUS_LABELS[ds]}
-                      </span>
-                    )
-                  })()}
-                </div>
-              </div>
-
-              <div className="flex-1 overflow-y-auto">
-                <div key={selectedComment.id} className="max-w-2xl mx-auto px-8 py-8 detail-enter">
-
-                  {selectedComment.screenshotUrl ? (
-                    <div className="rounded-xl border border-border overflow-hidden mb-6 bg-muted/40 flex items-center justify-center">
-                      <a
-                        href={selectedComment.screenshotUrl}
-                        target="_blank"
-                        rel="noopener noreferrer"
-                        className="block"
-                      >
-                        <img
-                          src={selectedComment.screenshotUrl}
-                          alt={`Screenshot of ${selectedComment.pageUrl}`}
-                          className="max-w-full max-h-[520px] w-auto h-auto object-contain"
-                          draggable={false}
-                        />
-                      </a>
-                    </div>
-                  ) : (
-                    <div className="rounded-xl border border-border bg-card p-5 mb-6">
-                      <div className="flex items-center gap-3 mb-3">
-                        <div className="w-8 h-8 rounded-lg bg-muted flex items-center justify-center">
-                          <ImageOffIcon size={16} className="text-muted-foreground" />
-                        </div>
-                        <div>
-                          <p className="text-xs font-semibold text-foreground">No screenshot captured</p>
-                          <p className="text-[11px] text-muted-foreground">Pin placed at ({selectedComment.x}, {selectedComment.y})</p>
-                        </div>
-                      </div>
-                      <div className="flex items-center gap-2 px-3 py-2 rounded-md bg-muted/60 border border-border">
-                        <SelectorIcon size={12} />
-                        <code className="text-[12px] font-mono text-foreground/70 break-all">{selectedComment.selector}</code>
-                      </div>
-                    </div>
-                  )}
-
-                  <div className="flex items-start gap-3 mb-8">
-                    <div
-                      className="w-8 h-8 rounded-full shrink-0 flex items-center justify-center text-xs font-bold text-white"
-                      style={{ background: selectedComment.authorColor }}
-                    >
-                      {selectedComment.authorInitial}
-                    </div>
-                    <div>
-                      <div className="flex items-center gap-2 mb-1">
-                        <span className="text-sm font-semibold text-foreground">{selectedComment.author}</span>
-                        <span className="text-xs text-muted-foreground">{timeAgo(selectedComment.createdAt)}</span>
-                      </div>
-                      <p className="text-[15px] leading-relaxed text-foreground">
-                        {selectedComment.body}
-                      </p>
-                      <div className="mt-2 text-xs font-mono text-muted-foreground">
-                        {selectedComment.selector}
-                      </div>
-                    </div>
-                  </div>
-                </div>
-              </div>
-
-              <div className="shrink-0 border-t border-border bg-card px-6 py-3">
-                <div className="flex items-center gap-2 max-w-2xl mx-auto">
-                  <ActionBtn
-                    active={selectedComment.reviewStatus === 'accepted' && selectedComment.implementationStatus !== 'done'}
-                    variant="accept"
-                    onClick={() => toggleReview(selectedComment, 'accepted')}
-                    shortcut="A"
-                  >
-                    <BotIcon size={14} /> Ready for Agent
-                  </ActionBtn>
-                  <ActionBtn
-                    active={selectedComment.implementationStatus === 'done'}
-                    variant="done"
-                    onClick={() => handleToggleDone(selectedComment.id)}
-                    shortcut="M"
-                  >
-                    <DoneIcon size={14} /> {selectedComment.implementationStatus === 'done' ? 'Done' : 'Mark Done'}
-                  </ActionBtn>
-                  <ActionBtn
-                    active={selectedComment.reviewStatus === 'rejected'}
-                    variant="reject"
-                    onClick={() => toggleReview(selectedComment, 'rejected')}
-                    shortcut="D"
-                  >
-                    <XIcon size={14} /> Reject
-                  </ActionBtn>
-
-                  <div className="w-px h-5 bg-border mx-1" />
-
-                  <ActionBtn variant="neutral" onClick={() => window.open(selectedComment.pageUrl, '_blank')}>
-                    <ExternalLinkIcon size={13} /> Open page
-                  </ActionBtn>
-
-                  <div className="flex-1" />
-
-                  <button
-                    onClick={goPrev}
-                    disabled={selectedIdx <= 0}
-                    className="p-1.5 rounded-md text-muted-foreground hover:text-foreground hover:bg-accent disabled:opacity-30 disabled:pointer-events-none transition-colors"
-                  >
-                    <ChevronLeftIcon size={16} />
-                  </button>
-                  <span className="text-xs font-mono text-muted-foreground tabular-nums">
-                    {selectedIdx + 1}/{filteredComments.length}
-                  </span>
-                  <button
-                    onClick={goNext}
-                    disabled={selectedIdx >= filteredComments.length - 1}
-                    className="p-1.5 rounded-md text-muted-foreground hover:text-foreground hover:bg-accent disabled:opacity-30 disabled:pointer-events-none transition-colors"
-                  >
-                    <ChevronRightIcon size={16} />
-                  </button>
-                </div>
-              </div>
-            </>
-          ) : selectedProject && !commentsLoading && !commentsError && projectComments.length === 0 ? (
-            <div className="flex-1 flex flex-col items-center justify-center text-center px-8">
-              <p className="text-base font-semibold text-foreground mb-1">No feedback yet</p>
-              <p className="text-sm text-muted-foreground max-w-sm mb-6">
-                Mount the widget in your app to start collecting visual feedback. Reviewers can drop pins on any page and their comments land here.
-              </p>
-              <AddToCodeButton projectId={selectedProject} apiBase={API_BASE} variant="large" />
-            </div>
-          ) : (
-            <div className="flex-1 flex flex-col items-center justify-center text-center px-8">
-              <div className="w-12 h-12 rounded-2xl bg-muted flex items-center justify-center mb-4">
-                <CursorIcon className="text-muted-foreground" />
-              </div>
-              <p className="text-base font-semibold text-foreground mb-1">Select a comment</p>
-              <p className="text-sm text-muted-foreground max-w-xs">
-                Pick a feedback item from the list to see the full context, screenshot, and actions.
-              </p>
-              <div className="flex gap-3 mt-6 text-xs text-muted-foreground font-mono">
-                <Kbd>J</Kbd><Kbd>K</Kbd> navigate
-                <span className="mx-1">·</span>
-                <Kbd>A</Kbd> ready
-                <span className="mx-1">·</span>
-                <Kbd>D</Kbd> reject
-              </div>
-            </div>
-          )}
-        </div>
-
-        {/* ── Right Panel: Agent Sidebar ── */}
         {sidebarOpen && (
-          <aside className="w-[300px] shrink-0 flex flex-col border-l border-border bg-sidebar overflow-y-auto animate-slide-in">
-            <div className="px-4 py-4 border-b border-sidebar-border">
-              <div className="flex items-center justify-between mb-1">
-                <h2 className="text-base font-bold text-foreground tracking-tight">Agent handoff</h2>
-                <button onClick={() => setSidebarOpen(false)} className="p-1.5 rounded-md text-muted-foreground hover:text-foreground hover:bg-accent transition-colors">
-                  <XIcon size={14} />
-                </button>
-              </div>
-              <p className="text-xs text-muted-foreground">Let AI agents fix your Ready for Agent items automatically</p>
-            </div>
-
-            {selectedProject && (
-              <div className="px-4 py-3 border-b border-sidebar-border">
-                <p className="text-[10px] font-bold text-muted-foreground uppercase tracking-wider mb-2">Install widget</p>
-                <AddToCodeButton projectId={selectedProject} apiBase={API_BASE} variant="compact" />
-                <p className="text-[10px] text-muted-foreground mt-2 leading-relaxed">
-                  Copies a setup prompt with package, mount snippet, and this project's credentials.
-                </p>
-              </div>
-            )}
-
-            {/* Session link — the core handoff */}
-            <div className="px-4 py-4 border-b border-sidebar-border animate-fade-in">
-              <p className="text-[10px] font-bold text-muted-foreground uppercase tracking-wider mb-2">Agent prompt</p>
-              <div className="rounded-lg border border-border bg-card p-3 mb-3">
-                <p className="text-[11px] font-mono text-foreground break-all leading-relaxed select-all">
-                  {agentError ? (
-                    <span className="text-status-rejected">{agentError}</span>
-                  ) : agentSession ? (
-                    formatDocUrl(agentSession.docUrl)
-                  ) : selectedProject ? (
-                    'Starting session…'
-                  ) : (
-                    'Select a project to start a session'
-                  )}
-                </p>
-              </div>
-              <button
-                onClick={handleCopySessionLink}
-                disabled={!agentSession}
-                className={cn(
-                  'w-full flex items-center justify-center gap-2 px-3 py-2 rounded-lg text-xs font-semibold transition-opacity btn-press',
-                  copyStatus === 'copied'
-                    ? 'bg-status-done text-white'
-                    : copyStatus === 'error'
-                    ? 'bg-status-rejected text-white'
-                    : 'bg-primary text-primary-foreground hover:opacity-90',
-                  !agentSession && 'opacity-40 pointer-events-none',
-                )}
-              >
-                <CopyIcon size={13} />
-                {copyStatus === 'copied' ? 'Copied ✓' : copyStatus === 'error' ? 'Copy failed' : `Copy ${selectedAgentMeta.name} prompt`}
-              </button>
-              <div className="mt-3 rounded-md bg-muted/60 border border-border px-3 py-2.5">
-                <p className="text-[11px] text-foreground font-medium mb-1.5">How it works</p>
-                <ol className="text-[10px] text-muted-foreground leading-relaxed space-y-1 list-decimal list-inside">
-                  <li>You review feedback and mark items as <span className="text-status-accepted font-semibold">Ready for Agent</span></li>
-                  <li>Copy this prompt and paste it into your AI agent</li>
-                  <li>The agent only sees <span className="text-status-accepted font-semibold">Ready for Agent</span> items — nothing else</li>
-                  <li>It claims, fixes, and marks them <span className="text-status-done font-semibold">Done</span></li>
-                </ol>
-              </div>
-            </div>
-
-            <div className="px-4 py-3 border-b border-sidebar-border">
-              <p className="text-[10px] font-bold text-muted-foreground uppercase tracking-wider mb-2">Your agent</p>
-              <div className="relative">
-                <button
-                  onClick={() => setAgentDropdownOpen((v) => !v)}
-                  className={cn(
-                    'w-full flex items-center justify-between px-3 py-2 rounded-lg border bg-card text-xs font-semibold transition-colors',
-                    agentDropdownOpen
-                      ? 'border-primary/40 text-foreground'
-                      : 'border-border text-foreground hover:border-muted-foreground/30'
-                  )}
-                >
-                  <div className="flex items-center gap-2">
-                    <BotIcon size={14} className="text-primary" />
-                    {AGENTS.find((a) => a.id === selectedAgent)?.name}
-                  </div>
-                  <ChevronDownIcon size={14} className={cn('text-muted-foreground transition-transform', agentDropdownOpen && 'rotate-180')} />
-                </button>
-                {agentDropdownOpen && (
-                  <>
-                    <button
-                      type="button"
-                      aria-label="Close agent selector"
-                      className="fixed inset-0 z-40 cursor-default"
-                      onClick={() => setAgentDropdownOpen(false)}
-                    />
-                    <div className="absolute top-full left-0 right-0 mt-1.5 rounded-lg border border-border bg-card shadow-xl shadow-black/30 z-50 py-1 cmd-modal-enter">
-                      {AGENTS.map((agent) => (
-                        <button
-                          key={agent.id}
-                          onClick={() => { setSelectedAgent(agent.id); setAgentDropdownOpen(false) }}
-                          className={cn(
-                            'w-full flex items-center justify-between px-3 py-2 text-left transition-colors',
-                            selectedAgent === agent.id
-                              ? 'bg-accent text-foreground'
-                              : 'text-foreground/80 hover:bg-accent/50'
-                          )}
-                        >
-                          <span className="text-xs font-semibold">{agent.name}</span>
-                          {selectedAgent === agent.id && <CheckIcon size={13} />}
-                        </button>
-                      ))}
-                    </div>
-                  </>
-                )}
-              </div>
-              <p className="text-[10px] text-muted-foreground mt-1.5">
-                {AGENTS.find((a) => a.id === selectedAgent)?.hint}
-              </p>
-            </div>
-
-            {agentConnected && agentShareState && (
-              <div className="px-4 py-3 border-b border-sidebar-border animate-fade-in">
-                <div className="flex items-center gap-2 mb-2.5">
-                  <div className="relative">
-                    <div className="w-2 h-2 rounded-full bg-agent-active animate-pulse-dot" />
-                    <div className="absolute inset-0 w-2 h-2 rounded-full animate-pulse-ring" />
-                  </div>
-                  <span className="text-[10px] font-bold text-agent-active uppercase tracking-wider">
-                    {agentShareState.presence.length === 1 ? 'Connected' : `${agentShareState.presence.length} connected`}
-                  </span>
-                </div>
-                <div className="flex flex-col gap-2">
-                  {agentShareState.presence.map((p) => (
-                    <div key={p.agentId} className="flex items-center gap-2.5 rounded-lg border border-border bg-card p-2.5 animate-scale-in">
-                      <div className="w-7 h-7 rounded-full bg-primary/10 flex items-center justify-center shrink-0">
-                        <BotIcon size={14} className="text-primary" />
-                      </div>
-                      <div className="flex-1 min-w-0">
-                        <p className="text-xs font-semibold text-foreground truncate">{p.agentId}</p>
-                        <p className="text-[11px] text-muted-foreground truncate">
-                          {p.summary ?? p.status}
-                        </p>
-                      </div>
-                      <div className="w-1.5 h-1.5 rounded-full bg-agent-active animate-pulse-dot shrink-0" />
-                    </div>
-                  ))}
-                </div>
-              </div>
-            )}
-
-            <div className="px-4 py-3 flex-1">
-              <p className="text-[10px] font-bold text-muted-foreground uppercase tracking-wider mb-3">Activity</p>
-              {agentEvents.length === 0 ? (
-                <p className="text-[11px] text-muted-foreground/70 leading-relaxed">
-                  {agentSession ? 'No agent activity yet. Paste the prompt into your agent to begin.' : 'Activity will appear once a session starts.'}
-                </p>
-              ) : (
-                <div className="space-y-0 animate-activity">
-                  {[...agentEvents].reverse().map((ev) => {
-                    const desc = describeEvent(ev)
-                    return (
-                      <div key={ev.id} className="flex gap-3 py-2 border-b border-border/40 last:border-0">
-                        <div className="mt-1.5 shrink-0">
-                          <div className={cn(
-                            'w-2 h-2 rounded-full',
-                            desc.kind === 'done' ? 'bg-status-accepted' :
-                            desc.kind === 'claim' ? 'bg-status-claimed' :
-                            desc.kind === 'file' ? 'bg-status-in-progress' :
-                            'bg-muted-foreground/30'
-                          )} />
-                        </div>
-                        <div className="flex-1 min-w-0">
-                          <p className="text-[12px] text-foreground leading-snug">{desc.text}</p>
-                          <p className="text-[10px] text-muted-foreground mt-0.5">{timeAgo(ev.createdAt)}</p>
-                        </div>
-                      </div>
-                    )
-                  })}
-                </div>
-              )}
-            </div>
-          </aside>
+          <AgentSidebar
+            apiBase={API_BASE}
+            selectedProject={selectedProject}
+            agentSession={agentSession}
+            agentShareState={agentShareState}
+            agentEvents={agentEvents}
+            agentError={agentError}
+            agentConnected={agentConnected}
+            selectedAgent={selectedAgent}
+            setSelectedAgent={setSelectedAgent}
+            selectedAgentMeta={selectedAgentMeta}
+            agentDropdownOpen={agentDropdownOpen}
+            setAgentDropdownOpen={setAgentDropdownOpen}
+            copyStatus={copyStatus}
+            onCopySessionLink={handleCopySessionLink}
+            onClose={() => setSidebarOpen(false)}
+          />
         )}
       </div>
 
-      {/* ── Status Bar ── */}
-      <footer className="flex items-center gap-5 px-5 h-[32px] shrink-0 border-t border-border bg-card text-[10px] font-mono text-muted-foreground">
-        <span><Kbd>A</Kbd> ready</span>
-        <span><Kbd>M</Kbd> done</span>
-        <span><Kbd>D</Kbd> reject</span>
-        <span><Kbd>Space</Kbd> next</span>
-        <span><Kbd>J</Kbd>/<Kbd>K</Kbd> nav</span>
-        <span><Kbd>S</Kbd> sidebar</span>
-        <span><Kbd>⌘K</Kbd> search</span>
-        <div className="flex-1" />
-        {!sidebarOpen && (
-          <button onClick={() => setSidebarOpen(true)} className="text-[11px] font-semibold text-primary hover:underline">
-            Show agent panel
-          </button>
-        )}
-      </footer>
+      <StatusBar sidebarOpen={sidebarOpen} onShowSidebar={() => setSidebarOpen(true)} />
 
       {cmdOpen && (
         <CommandPalette
@@ -1118,560 +361,5 @@ export function App() {
         />
       )}
     </div>
-  )
-}
-
-// ─── Command Palette ────────────────────────────────────────────────
-
-type CmdIconType = 'comment' | 'check' | 'x' | 'filter' | 'sidebar'
-type CmdItem = { id: string; type: 'comment' | 'action'; label: string; detail: string; icon: CmdIconType }
-
-const CMD_ACTIONS: CmdItem[] = [
-  { id: 'accept', type: 'action', label: 'Toggle Ready for Agent', detail: 'A', icon: 'check' },
-  { id: 'done', type: 'action', label: 'Toggle Done', detail: 'M', icon: 'check' },
-  { id: 'reject', type: 'action', label: 'Toggle Reject', detail: 'D', icon: 'x' },
-  { id: 'toggle-sidebar', type: 'action', label: 'Toggle agent panel', detail: 'S', icon: 'sidebar' },
-  { id: 'filter-all', type: 'action', label: 'Filter: All', detail: '', icon: 'filter' },
-  { id: 'filter-open', type: 'action', label: 'Filter: Open', detail: '', icon: 'filter' },
-  { id: 'filter-ready', type: 'action', label: 'Filter: Ready for Agent', detail: '', icon: 'filter' },
-  { id: 'filter-done', type: 'action', label: 'Filter: Done', detail: '', icon: 'filter' },
-]
-
-interface CommandPaletteProps {
-  onClose: () => void
-  comments: Comment[]
-  onSelect: (commentId: string) => void
-  onAction: (action: string) => void
-  selectedCommentId: string
-}
-
-function CommandPalette({ onClose, comments, onSelect, onAction, selectedCommentId }: CommandPaletteProps) {
-  const [query, setQuery] = useState('')
-  const [activeIdx, setActiveIdx] = useState(0)
-  const inputRef = useRef<HTMLInputElement>(null)
-  const listRef = useRef<HTMLDivElement>(null)
-
-  useEffect(() => {
-    const h = requestAnimationFrame(() => inputRef.current?.focus())
-    return () => cancelAnimationFrame(h)
-  }, [])
-
-  const results = useMemo<CmdItem[]>(() => {
-    const q = query.toLowerCase().trim()
-
-    const matchedComments: CmdItem[] = comments
-      .filter((c) => {
-        if (!q) return true
-        return (
-          c.body.toLowerCase().includes(q) ||
-          c.author.toLowerCase().includes(q) ||
-          c.selector.toLowerCase().includes(q) ||
-          c.pageUrl.toLowerCase().includes(q)
-        )
-      })
-      .slice(0, 8)
-      .map((c) => ({
-        id: c.id,
-        type: 'comment',
-        label: c.body.length > 80 ? c.body.slice(0, 80) + '…' : c.body,
-        detail: `${c.author} · ${truncateUrl(c.pageUrl)}`,
-        icon: 'comment',
-      }))
-
-    const matchedActions = q
-      ? CMD_ACTIONS.filter((a) => a.label.toLowerCase().includes(q))
-      : CMD_ACTIONS
-
-    return q ? [...matchedComments, ...matchedActions] : [...matchedActions, ...matchedComments]
-  }, [query, comments])
-
-  useEffect(() => {
-    const list = listRef.current
-    if (!list) return
-    const el = list.children[activeIdx] as HTMLElement | undefined
-    el?.scrollIntoView({ block: 'nearest' })
-  }, [activeIdx])
-
-  const handleSelect = useCallback(() => {
-    const item = results[activeIdx]
-    if (!item) return
-    if (item.type === 'comment') onSelect(item.id)
-    else onAction(item.id)
-  }, [results, activeIdx, onSelect, onAction])
-
-  const handleKeyDown = useCallback((e: React.KeyboardEvent) => {
-    if (e.key === 'ArrowDown') {
-      e.preventDefault()
-      setActiveIdx((i) => Math.min(i + 1, results.length - 1))
-    } else if (e.key === 'ArrowUp') {
-      e.preventDefault()
-      setActiveIdx((i) => Math.max(i - 1, 0))
-    } else if (e.key === 'Enter') {
-      e.preventDefault()
-      handleSelect()
-    } else if (e.key === 'Escape') {
-      onClose()
-    }
-  }, [results.length, handleSelect, onClose])
-
-  return (
-    <div className="fixed inset-0 z-50 flex justify-center pt-[20vh]">
-      <button
-        type="button"
-        aria-label="Close command palette"
-        className="absolute inset-0 bg-background/60 backdrop-blur-sm cmd-backdrop-enter"
-        onClick={onClose}
-      />
-      <div className="relative w-full max-w-[540px] h-fit rounded-xl border border-border bg-card shadow-2xl shadow-black/40 overflow-hidden cmd-modal-enter">
-        <div className="flex items-center gap-3 px-4 h-[52px] border-b border-border">
-          <SearchIcon />
-          <input
-            ref={inputRef}
-            type="text"
-            value={query}
-            onChange={(e) => {
-              setQuery(e.target.value)
-              setActiveIdx(0)
-            }}
-            onKeyDown={handleKeyDown}
-            placeholder="Search feedback, jump to comment, or run action…"
-            className="flex-1 bg-transparent text-sm text-foreground placeholder:text-muted-foreground outline-none"
-            autoComplete="off"
-            spellCheck={false}
-          />
-          <kbd className="px-1.5 py-0.5 rounded border border-border bg-muted text-[10px] font-mono text-muted-foreground">
-            ESC
-          </kbd>
-        </div>
-
-        <div ref={listRef} className="max-h-[360px] overflow-y-auto py-2">
-          {results.length === 0 ? (
-            <div className="px-4 py-8 text-center">
-              <p className="text-sm text-muted-foreground">No results for "{query}"</p>
-            </div>
-          ) : (
-            results.map((item, i) => {
-              const prevType = results[i - 1]?.type
-              const showHeader = i === 0 || item.type !== prevType
-
-              return (
-                <div key={`${item.type}-${item.id}`}>
-                  {showHeader && (
-                    <div className="px-4 pt-2 pb-1">
-                      <span className="text-[10px] font-bold text-muted-foreground uppercase tracking-wider">
-                        {item.type === 'action' ? 'Actions' : 'Comments'}
-                      </span>
-                    </div>
-                  )}
-                  <button
-                    className={cn(
-                      'w-full flex items-center gap-3 px-4 py-2.5 text-left transition-colors',
-                      i === activeIdx
-                        ? 'bg-accent text-foreground'
-                        : 'text-foreground/80 hover:bg-accent/50'
-                    )}
-                    onMouseEnter={() => setActiveIdx(i)}
-                    onClick={() => {
-                      if (item.type === 'comment') onSelect(item.id)
-                      else onAction(item.id)
-                    }}
-                  >
-                    <div className={cn(
-                      'w-7 h-7 rounded-lg shrink-0 flex items-center justify-center',
-                      i === activeIdx ? 'bg-primary/15 text-primary' : 'bg-muted text-muted-foreground'
-                    )}>
-                      <CmdIcon type={item.icon} />
-                    </div>
-
-                    <div className="flex-1 min-w-0">
-                      <p className={cn(
-                        'text-[13px] leading-snug truncate',
-                        i === activeIdx ? 'text-foreground' : 'text-foreground/80',
-                        item.type === 'comment' && item.id === selectedCommentId && 'font-semibold'
-                      )}>
-                        {item.label}
-                      </p>
-                      {item.detail && (
-                        <p className="text-[11px] text-muted-foreground truncate mt-0.5">
-                          {item.detail}
-                        </p>
-                      )}
-                    </div>
-
-                    {item.type === 'action' && item.detail ? (
-                      <kbd className="px-1.5 py-0.5 rounded border border-border bg-muted text-[10px] font-mono text-muted-foreground shrink-0">
-                        {item.detail}
-                      </kbd>
-                    ) : (
-                      i === activeIdx && (
-                        <span className="text-muted-foreground shrink-0">
-                          <ReturnIcon />
-                        </span>
-                      )
-                    )}
-                  </button>
-                </div>
-              )
-            })
-          )}
-        </div>
-
-        <div className="flex items-center gap-4 px-4 h-[36px] border-t border-border text-[10px] font-mono text-muted-foreground">
-          <span className="flex items-center gap-1"><Kbd>↑</Kbd><Kbd>↓</Kbd> navigate</span>
-          <span className="flex items-center gap-1"><Kbd>↵</Kbd> select</span>
-          <span className="flex items-center gap-1"><Kbd>esc</Kbd> close</span>
-        </div>
-      </div>
-    </div>
-  )
-}
-
-function AddProjectPopover({ onAdd, onClose, busy, error }: {
-  onAdd: (name: string) => void
-  onClose: () => void
-  busy?: boolean
-  error?: string | null
-}) {
-  const [name, setName] = useState('')
-  const inputRef = useRef<HTMLInputElement>(null)
-
-  useEffect(() => {
-    requestAnimationFrame(() => inputRef.current?.focus())
-
-    function handleClick(e: MouseEvent) {
-      const target = e.target as HTMLElement
-      if (!target.closest('[data-add-project]')) onClose()
-    }
-    function handleKey(e: KeyboardEvent) {
-      if (e.key === 'Escape') onClose()
-    }
-    document.addEventListener('mousedown', handleClick)
-    document.addEventListener('keydown', handleKey)
-    return () => {
-      document.removeEventListener('mousedown', handleClick)
-      document.removeEventListener('keydown', handleKey)
-    }
-  }, [onClose])
-
-  const canSubmit = !!name.trim() && !busy
-
-  return (
-    <div
-      data-add-project
-      className="absolute top-full left-0 mt-2 w-[260px] rounded-lg border border-border bg-card shadow-xl shadow-black/30 cmd-modal-enter z-50"
-    >
-      <div className="p-3">
-        <p className="text-[11px] font-bold text-muted-foreground uppercase tracking-wider mb-2">New project</p>
-        <form
-          onSubmit={(e) => {
-            e.preventDefault()
-            if (canSubmit) onAdd(name.trim())
-          }}
-        >
-          <input
-            ref={inputRef}
-            type="text"
-            value={name}
-            onChange={(e) => setName(e.target.value)}
-            placeholder="Project name"
-            disabled={busy}
-            className="w-full px-3 py-2 rounded-md border border-border bg-background text-sm text-foreground placeholder:text-muted-foreground outline-none focus:border-primary/50 focus:ring-1 focus:ring-primary/20 transition-colors disabled:opacity-50"
-            autoComplete="off"
-            spellCheck={false}
-          />
-          {error && (
-            <p className="mt-2 text-[11px] text-status-rejected">{error}</p>
-          )}
-          <div className="flex items-center justify-between mt-3">
-            <p className="text-[10px] text-muted-foreground">
-              Paste your snippet after creating
-            </p>
-            <button
-              type="submit"
-              disabled={!canSubmit}
-              className={cn(
-                'px-3 py-1.5 rounded-md text-xs font-semibold transition-all btn-press',
-                canSubmit
-                  ? 'bg-primary text-primary-foreground hover:opacity-90'
-                  : 'bg-muted text-muted-foreground cursor-not-allowed'
-              )}
-            >
-              {busy ? 'Creating…' : 'Create'}
-            </button>
-          </div>
-        </form>
-      </div>
-    </div>
-  )
-}
-
-function CmdIcon({ type }: { type: CmdIconType }) {
-  switch (type) {
-    case 'comment': return <ChatIcon />
-    case 'check': return <CheckIcon size={14} />
-    case 'x': return <XIcon size={14} />
-    case 'filter': return <SvgIcon d="M22 3H2l8 9.46V19l4 2v-8.54L22 3z" size={14} />
-    case 'sidebar': return <SvgIcon d="M21 3H3a2 2 0 00-2 2v14a2 2 0 002 2h18a2 2 0 002-2V5a2 2 0 00-2-2zM15 3v18" size={14} />
-  }
-}
-
-function ReturnIcon() {
-  return (
-    <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
-      <polyline points="9 10 4 15 9 20" /><path d="M20 4v7a4 4 0 01-4 4H4" />
-    </svg>
-  )
-}
-
-// ─── Small Components ───────────────────────────────────────────────
-
-function StatusBadge({ comment }: { comment: Comment }) {
-  const ds = getDisplayStatus(comment)
-  return (
-    <span className={cn(
-      'inline-flex items-center px-2 py-0.5 rounded text-[10px] font-semibold',
-      ds === 'ready' && 'bg-status-accepted-bg text-status-accepted',
-      ds === 'rejected' && 'bg-status-rejected-bg text-status-rejected',
-      ds === 'open' && 'bg-status-open-bg text-status-open',
-      ds === 'done' && 'bg-status-done-bg text-status-done',
-    )}>
-      {DISPLAY_STATUS_LABELS[ds]}
-    </span>
-  )
-}
-
-function ActionBtn({ children, variant, active, onClick, shortcut, disabled }: {
-  children: React.ReactNode
-  variant: 'accept' | 'reject' | 'neutral' | 'done'
-  active?: boolean
-  onClick?: () => void
-  shortcut?: string
-  disabled?: boolean
-}) {
-  return (
-    <button
-      type="button"
-      onClick={onClick}
-      disabled={disabled}
-      title={shortcut ? `${shortcut}` : undefined}
-      className={cn(
-        'inline-flex items-center gap-1.5 px-3 py-1.5 rounded-lg text-xs font-semibold transition-all',
-        'disabled:opacity-40 disabled:pointer-events-none btn-press',
-        variant === 'accept' && (active
-          ? 'bg-status-accepted-bg text-status-accepted border border-status-accepted/30'
-          : 'border border-border bg-card text-foreground hover:bg-status-accepted-bg hover:text-status-accepted hover:border-status-accepted/30'
-        ),
-        variant === 'reject' && (active
-          ? 'bg-status-rejected-bg text-status-rejected border border-status-rejected/30'
-          : 'border border-border bg-card text-foreground hover:bg-status-rejected-bg hover:text-status-rejected hover:border-status-rejected/30'
-        ),
-        variant === 'done' && (active
-          ? 'bg-status-done-bg text-status-done border border-status-done/30'
-          : 'border border-border bg-card text-foreground hover:bg-status-done-bg hover:text-status-done hover:border-status-done/30'
-        ),
-        variant === 'neutral' && 'border border-border bg-card text-foreground hover:bg-accent',
-      )}
-    >
-      {children}
-    </button>
-  )
-}
-
-function AddToCodeButton({ projectId, apiBase, variant = 'compact' }: {
-  projectId: string
-  apiBase: string
-  variant?: 'compact' | 'large'
-}) {
-  const [copied, setCopied] = useState(false)
-  const copiedTimerRef = useRef<ReturnType<typeof setTimeout> | null>(null)
-
-  useEffect(() => () => {
-    if (copiedTimerRef.current) clearTimeout(copiedTimerRef.current)
-  }, [])
-
-  const handleCopy = useCallback(async () => {
-    const prompt = buildIntegrationPrompt(projectId, apiBase)
-    try {
-      await navigator.clipboard.writeText(prompt)
-      setCopied(true)
-      if (copiedTimerRef.current) clearTimeout(copiedTimerRef.current)
-      copiedTimerRef.current = setTimeout(() => setCopied(false), 2000)
-    } catch (err) {
-      console.error('Clipboard write failed:', err)
-    }
-  }, [projectId, apiBase])
-
-  if (variant === 'large') {
-    return (
-      <div className="w-full max-w-md mx-auto rounded-2xl border border-border bg-card p-6 text-left">
-        <div className="flex items-center gap-2.5 mb-1">
-          <div className="w-7 h-7 rounded-lg bg-primary/10 flex items-center justify-center">
-            <BotIcon size={15} className="text-primary" />
-          </div>
-          <p className="text-sm font-bold text-foreground">Add widget to your code</p>
-        </div>
-        <p className="text-xs text-muted-foreground mb-4 leading-relaxed">
-          Copy a setup prompt for any AI coding agent. It includes the npm package, mount snippet, and this project's credentials.
-        </p>
-        <div className="rounded-md bg-muted/60 border border-border px-3 py-2 mb-4 space-y-1">
-          <div className="flex items-center gap-2 text-[11px] font-mono text-muted-foreground">
-            <span className="opacity-60 w-[60px] shrink-0">project</span>
-            <span className="text-foreground truncate">{projectId}</span>
-          </div>
-          <div className="flex items-center gap-2 text-[11px] font-mono text-muted-foreground">
-            <span className="opacity-60 w-[60px] shrink-0">apiBase</span>
-            <span className="text-foreground truncate">{apiBase}</span>
-          </div>
-        </div>
-        <button
-          onClick={handleCopy}
-          className="w-full flex items-center justify-center gap-2 px-3 py-2 rounded-lg bg-primary text-primary-foreground text-xs font-semibold hover:opacity-90 transition-opacity btn-press"
-        >
-          {copied ? <CheckIcon size={13} /> : <CopyIcon size={13} />}
-          {copied ? 'Copied to clipboard' : 'Copy install prompt'}
-        </button>
-        <p className="text-[10px] text-muted-foreground mt-3 leading-relaxed">
-          Paste it into your AI coding agent inside the project repo.
-        </p>
-      </div>
-    )
-  }
-
-  return (
-    <button
-      onClick={handleCopy}
-      className="w-full flex items-center justify-center gap-2 px-3 py-2 rounded-lg border border-border bg-card text-foreground text-xs font-semibold hover:bg-accent transition-colors btn-press"
-      title="Copy a setup prompt for your AI coding agent with this project's credentials"
-    >
-      {copied ? <CheckIcon size={13} /> : <CopyIcon size={13} />}
-      {copied ? 'Copied' : 'Add to code'}
-    </button>
-  )
-}
-
-function Spinner({ size = 20, strokeWidth = 2.5, className = 'text-muted-foreground' }: {
-  size?: number
-  strokeWidth?: number
-  className?: string
-}) {
-  return (
-    <svg width={size} height={size} viewBox="0 0 24 24" fill="none" className={cn('animate-spin', className)}>
-      <circle cx="12" cy="12" r="9" stroke="currentColor" strokeWidth={strokeWidth} opacity="0.25" />
-      <path d="M21 12a9 9 0 0 0-9-9" stroke="currentColor" strokeWidth={strokeWidth} strokeLinecap="round" />
-    </svg>
-  )
-}
-
-function Kbd({ children }: { children: React.ReactNode }) {
-  return (
-    <kbd className="inline-flex items-center justify-center min-w-[18px] px-1 py-0.5 rounded border border-border bg-muted text-[9px] font-mono font-semibold text-muted-foreground">
-      {children}
-    </kbd>
-  )
-}
-
-// ─── Icons ──────────────────────────────────────────────────────────
-
-function SvgIcon({ d, size = 16, className }: { d: string; size?: number; className?: string }) {
-  return (
-    <svg width={size} height={size} viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" className={className}>
-      <path d={d} />
-    </svg>
-  )
-}
-
-function CheckIcon({ size = 16 }: { size?: number }) {
-  return <SvgIcon d="M5 12l5 5L20 7" size={size} />
-}
-
-function XIcon({ size = 16 }: { size?: number }) {
-  return (
-    <svg width={size} height={size} viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
-      <line x1="6" y1="6" x2="18" y2="18" /><line x1="18" y1="6" x2="6" y2="18" />
-    </svg>
-  )
-}
-
-function ExternalLinkIcon({ size = 16 }: { size?: number }) {
-  return <SvgIcon d="M18 13v6a2 2 0 01-2 2H5a2 2 0 01-2-2V8a2 2 0 012-2h6M15 3h6v6M10 14L21 3" size={size} />
-}
-
-function ChevronLeftIcon({ size = 16 }: { size?: number }) {
-  return <SvgIcon d="M15 18l-6-6 6-6" size={size} />
-}
-
-function ChevronRightIcon({ size = 16 }: { size?: number }) {
-  return <SvgIcon d="M9 18l6-6-6-6" size={size} />
-}
-
-function CopyIcon({ size = 16 }: { size?: number }) {
-  return (
-    <svg width={size} height={size} viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
-      <rect x="9" y="9" width="13" height="13" rx="2" ry="2" /><path d="M5 15H4a2 2 0 01-2-2V4a2 2 0 012-2h9a2 2 0 012 2v1" />
-    </svg>
-  )
-}
-
-function BotIcon({ size = 16, className }: { size?: number; className?: string }) {
-  return (
-    <svg width={size} height={size} viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" className={className}>
-      <path d="M12 8V4H8" /><rect x="4" y="8" width="16" height="12" rx="2" /><path d="M2 14h2M20 14h2M15 13v2M9 13v2" />
-    </svg>
-  )
-}
-
-function DoneIcon({ size = 16 }: { size?: number }) {
-  return (
-    <svg width={size} height={size} viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
-      <path d="M22 11.08V12a10 10 0 11-5.93-9.14" /><polyline points="22 4 12 14.01 9 11.01" />
-    </svg>
-  )
-}
-
-function SelectorIcon({ size = 16 }: { size?: number }) {
-  return <SvgIcon d="M7 2h10M7 22h10M2 7v10M22 7v10" size={size} />
-}
-
-function ImageOffIcon({ size = 16, className }: { size?: number; className?: string }) {
-  return (
-    <svg width={size} height={size} viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" className={className}>
-      <path d="M2 2l20 20M9 9a3 3 0 104.24 4.24" /><path d="M21 15V5a2 2 0 00-2-2H9" /><path d="M3 9v10a2 2 0 002 2h10l3.3-3.3" /><path d="M14 14l2.44-2.44" />
-    </svg>
-  )
-}
-
-function ChevronDownIcon({ size = 16, className }: { size?: number; className?: string }) {
-  return <SvgIcon d="M6 9l6 6 6-6" size={size} className={className} />
-}
-
-function PlusIcon({ size = 16 }: { size?: number }) {
-  return (
-    <svg width={size} height={size} viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
-      <line x1="12" y1="5" x2="12" y2="19" /><line x1="5" y1="12" x2="19" y2="12" />
-    </svg>
-  )
-}
-
-function SearchIcon() {
-  return <SvgIcon d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z" size={14} />
-}
-
-function CheckboxIcon() {
-  return (
-    <svg width="14" height="14" viewBox="0 0 16 16" fill="none" stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round">
-      <rect x="2" y="2" width="12" height="12" rx="2" />
-    </svg>
-  )
-}
-
-function ChatIcon({ className }: { className?: string }) {
-  return <SvgIcon d="M21 15a2 2 0 01-2 2H7l-4 4V5a2 2 0 012-2h14a2 2 0 012 2v10z" size={20} className={className} />
-}
-
-function CursorIcon({ className }: { className?: string }) {
-  return (
-    <svg width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" className={className}>
-      <path d="M5 3l14 9-7 2-3 7L5 3z" />
-    </svg>
   )
 }
