+++ apps/dashboard/components/icons.tsx
@@ -1,0 +1,148 @@
+export function SvgIcon({ d, size = 16, className }: { d: string; size?: number; className?: string }) {
+  return (
+    <svg width={size} height={size} viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" className={className}>
+      <path d={d} />
+    </svg>
+  )
+}
+
+export function CheckIcon({ size = 16 }: { size?: number }) {
+  return <SvgIcon d="M5 12l5 5L20 7" size={size} />
+}
+
+export function XIcon({ size = 16 }: { size?: number }) {
+  return (
+    <svg width={size} height={size} viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
+      <line x1="6" y1="6" x2="18" y2="18" /><line x1="18" y1="6" x2="6" y2="18" />
+    </svg>
+  )
+}
+
+export function ExternalLinkIcon({ size = 16 }: { size?: number }) {
+  return <SvgIcon d="M18 13v6a2 2 0 01-2 2H5a2 2 0 01-2-2V8a2 2 0 012-2h6M15 3h6v6M10 14L21 3" size={size} />
+}
+
+export function ChevronLeftIcon({ size = 16 }: { size?: number }) {
+  return <SvgIcon d="M15 18l-6-6 6-6" size={size} />
+}
+
+export function ChevronRightIcon({ size = 16 }: { size?: number }) {
+  return <SvgIcon d="M9 18l6-6-6-6" size={size} />
+}
+
+export function CopyIcon({ size = 16 }: { size?: number }) {
+  return (
+    <svg width={size} height={size} viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
+      <rect x="9" y="9" width="13" height="13" rx="2" ry="2" /><path d="M5 15H4a2 2 0 01-2-2V4a2 2 0 012-2h9a2 2 0 012 2v1" />
+    </svg>
+  )
+}
+
+export function BotIcon({ size = 16, className }: { size?: number; className?: string }) {
+  return (
+    <svg width={size} height={size} viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" className={className}>
+      <path d="M12 8V4H8" /><rect x="4" y="8" width="16" height="12" rx="2" /><path d="M2 14h2M20 14h2M15 13v2M9 13v2" />
+    </svg>
+  )
+}
+
+export function DoneIcon({ size = 16 }: { size?: number }) {
+  return (
+    <svg width={size} height={size} viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
+      <path d="M22 11.08V12a10 10 0 11-5.93-9.14" /><polyline points="22 4 12 14.01 9 11.01" />
+    </svg>
+  )
+}
+
+export function SelectorIcon({ size = 16 }: { size?: number }) {
+  return <SvgIcon d="M7 2h10M7 22h10M2 7v10M22 7v10" size={size} />
+}
+
+export function ImageOffIcon({ size = 16, className }: { size?: number; className?: string }) {
+  return (
+    <svg width={size} height={size} viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" className={className}>
+      <path d="M2 2l20 20M9 9a3 3 0 104.24 4.24" /><path d="M21 15V5a2 2 0 00-2-2H9" /><path d="M3 9v10a2 2 0 002 2h10l3.3-3.3" /><path d="M14 14l2.44-2.44" />
+    </svg>
+  )
+}
+
+export function ChevronDownIcon({ size = 16, className }: { size?: number; className?: string }) {
+  return <SvgIcon d="M6 9l6 6 6-6" size={size} className={className} />
+}
+
+export function PlusIcon({ size = 16 }: { size?: number }) {
+  return (
+    <svg width={size} height={size} viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
+      <line x1="12" y1="5" x2="12" y2="19" /><line x1="5" y1="12" x2="19" y2="12" />
+    </svg>
+  )
+}
+
+export function SearchIcon() {
+  return <SvgIcon d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z" size={14} />
+}
+
+export function CheckboxIcon() {
+  return (
+    <svg width="14" height="14" viewBox="0 0 16 16" fill="none" stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round">
+      <rect x="2" y="2" width="12" height="12" rx="2" />
+    </svg>
+  )
+}
+
+export function ChatIcon({ className }: { className?: string }) {
+  return <SvgIcon d="M21 15a2 2 0 01-2 2H7l-4 4V5a2 2 0 012-2h14a2 2 0 012 2v10z" size={20} className={className} />
+}
+
+export function CursorIcon({ className }: { className?: string }) {
+  return (
+    <svg width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" className={className}>
+      <path d="M5 3l14 9-7 2-3 7L5 3z" />
+    </svg>
+  )
+}
+
+export function ReturnIcon() {
+  return (
+    <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
+      <polyline points="9 10 4 15 9 20" /><path d="M20 4v7a4 4 0 01-4 4H4" />
+    </svg>
+  )
+}
+
+export type CmdIconType = 'comment' | 'check' | 'x' | 'filter' | 'sidebar'
+
+export function CmdIcon({ type }: { type: CmdIconType }) {
+  switch (type) {
+    case 'comment': return <ChatIcon />
+    case 'check': return <CheckIcon size={14} />
+    case 'x': return <XIcon size={14} />
+    case 'filter': return <SvgIcon d="M22 3H2l8 9.46V19l4 2v-8.54L22 3z" size={14} />
+    case 'sidebar': return <SvgIcon d="M21 3H3a2 2 0 00-2 2v14a2 2 0 002 2h18a2 2 0 002-2V5a2 2 0 00-2-2zM15 3v18" size={14} />
+  }
+}
+
+export function LogoIcon() {
+  return (
+    <svg width="13" height="13" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2.5" strokeLinecap="round" strokeLinejoin="round" className="text-primary-foreground">
+      <path d="M21 15a2 2 0 01-2 2H7l-4 4V5a2 2 0 012-2h14a2 2 0 012 2v10z" />
+    </svg>
+  )
+}
+
+export function MoonIcon({ size = 14 }: { size?: number }) {
+  return (
+    <svg width={size} height={size} viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
+      <path d="M21 12.79A9 9 0 1 1 11.21 3 7 7 0 0 0 21 12.79z" />
+    </svg>
+  )
+}
+
+export function SunIcon({ size = 14 }: { size?: number }) {
+  return (
+    <svg width={size} height={size} viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round">
+      <circle cx="12" cy="12" r="4" />
+      <path d="M12 2v2M12 20v2M4.93 4.93l1.41 1.41M17.66 17.66l1.41 1.41M2 12h2M20 12h2M4.93 19.07l1.41-1.41M17.66 6.34l1.41-1.41" />
+    </svg>
+  )
+}
